+++ CBF.Domain/Entities/EstatisticaJogadorClube.cs
@@ -9,8 +9,8 @@ namespace CBF.Domain.Entities
 {
     public class EstatisticaJogadorClube : EntityBase
     {
-        public int IdJogador { get; set; }
-        public int IdClube { get; set; }
+        public long IdJogador { get; set; }
+        public long IdClube { get; set; }
         public int Ano { get; set; }
         public int Partidas { get; set; }
         public int Gols { get; set; }
